+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/ ), and followin
 
 ## [Unreleased][unreleased]
 ### Added
- 
+ - added an option to teleport to a chunkloader (if you have the permission)
+
 ### Changed
  -  the info command is now only showing active currently loaded chunks.